move apiError to https package
